[Refactor] gps 검사를 HomeScreen에서
