+++ app/src/main/java/com/and04/naturealbum/ui/NatureAlbumApp.kt
@@ -7,7 +7,6 @@ import android.content.Intent
 import android.net.Uri
 import android.provider.MediaStore
 import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.IntentSenderRequest
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
@@ -17,7 +16,6 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.and04.naturealbum.ui.home.GPSHandler
 import com.and04.naturealbum.ui.home.HomeScreen
 import com.and04.naturealbum.ui.savephoto.SavePhotoScreen
 import com.and04.naturealbum.ui.theme.NatureAlbumTheme
@@ -62,37 +60,15 @@ fun NatureAlbumNavHost(
             }
         }
     }
-    val locationSettingsLauncher =
-        rememberLauncherForActivityResult(
-            contract = ActivityResultContracts.StartIntentSenderForResult()
-        ) { result ->
-            if (result.resultCode == RESULT_OK) {
-                takePicture()
-            }
-        }
-    val isGrantedGPS = { intentSenderRequest: IntentSenderRequest ->
-        locationSettingsLauncher.launch(intentSenderRequest)
-    }
-    val allPermissionGranted = {
-        GPSHandler(
-            activity,
-            { intentSenderRequest ->
-                isGrantedGPS(
-                    intentSenderRequest
-                )
-            },
-            {
-                takePicture()
-            }
-        ).startLocationUpdates()
-    }
+
+
 
     NavHost(
         navController = navController,
         startDestination = NavigateDestination.Home.route
     ) {
         composable(NavigateDestination.Home.route) {
-            HomeScreen(allPermissionGranted = { allPermissionGranted() })
+            HomeScreen(takePicture = { takePicture() })
         }
         composable(NavigateDestination.SavePhoto.route) {
             imageUri?.let { uri -> SavePhotoScreen(model = uri) }
